Creating a XmlRpc Manager to handle blog submissions via live writer.
