+++ source/DasBlog.Web.Repositories/Interfaces/IBlogManager.cs
@@ -31,7 +31,5 @@ namespace DasBlog.Managers.Interfaces
 		CommentSaveState ApproveComment(string postid, string commentid);
 
 		CommentCollection GetComments(string postid, bool allComments);
-
-		string XmlRpcInvoke(Stream blob);
 	}
 }
